requirements.txt
PyYAML>=6.0
fastapi>=0.115.0
uvicorn>=0.30.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
requests>=2.31.0
lark-oapi>=1.4.10
loguru>=0.7.0
SQLAlchemy>=2.0
psycopg[binary]>=3.2
alembic>=1.13
pytest>=8.0
